requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0
httpx>=0.24.0,<1.0.0
pydantic>=2.0.1,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
structlog>=23.1.0,<26.0.0
python-dateutil>=2.8.2,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pandas>=1.5.0,<3.0.0
numpy>=1.23.2,<2.0.0
sqlglot>=18.0.0,<30.0.0
yahooquery>=2.3.0,<3.0.0
pandas-datareader>=0.10.0,<1.0.0
pyfredapi>=0.6.0,<1.0.0
yfinance>=0.2.36,<1.0.0
# Optional (heavy); leave installed if you plan to use it:
openbb>=4.5.0,<5.0.0
